% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accgrad.R
\name{accgrad}
\alias{accgrad}
\title{Sub-function \code{accgrad()} for tree-lasso}
\usage{
accgrad(
  y,
  x,
  lambda,
  Tree,
  C,
  g_idx,
  TauNorm,
  mu,
  option,
  num_nonpen = 0,
  intercept = TRUE,
  y_mis = NULL,
  x_mis = NULL
)
}
\description{
Based on the matlab code from http://www.cs.cmu.edu/~sssykim/softwares/softwares.html
}
\details{
mixlasso
}
