% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accgrad.re.R
\name{accgrad.re}
\alias{accgrad.re}
\title{Sub-function \code{accgrad()} for tree-lasso}
\usage{
accgrad.re(
  y,
  x,
  lambda,
  Tree,
  t.idx,
  C,
  g_idx,
  TauNorm,
  mu,
  NoVar,
  option,
  num_nonpen = 0,
  intercept = TRUE,
  y_mis = NULL,
  x_mis = NULL,
  t.glasso = FALSE,
  alpha = 1,
  gamma = 0,
  cov.proxy = "FL",
  L = 0,
  predict.re = FALSE
)
}
\description{
A version of \code{accgrad()} with random effects for clustered samples
}
\details{
mixlasso
}
