% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convH2T.R
\name{convH2T}
\alias{convH2T}
\title{Sub-function \code{covH2T()} for tree-lasso}
\usage{
convH2T(H, w_max)
}
\description{
Based on the matlab code from http://www.cs.cmu.edu/~sssykim/softwares/softwares.html
}
\details{
mixlasso
}
