% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convNd2T.R
\name{convNd2T}
\alias{convNd2T}
\title{Sub-function \code{convNd2T()} for tree-lasso}
\usage{
convNd2T(Nd, w, w_max)
}
\description{
Based on the matlab code from http://www.cs.cmu.edu/~sssykim/softwares/softwares.html
}
\details{
mixlasso
}
