% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCorr.R
\name{fastCorr}
\alias{fastCorr}
\title{Sub-function \code{fastCorr()} for tree-lasso}
\usage{
fastCorr(A)
}
\description{
Based on the matlab code from http://www.cs.cmu.edu/~sssykim/softwares/softwares.html
}
\details{
mixlasso
}
