% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.grad.R
\name{pre.grad}
\alias{pre.grad}
\title{Sub-function \code{pre.grad()} for tree-lasso}
\usage{
pre.grad(Tree, Tw)
}
\description{
Based on the matlab code from http://www.cs.cmu.edu/~sssykim/softwares/softwares.html
}
\details{
mixlasso
}
