% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.lasso.R
\name{tree.lasso}
\alias{tree.lasso}
\title{Sub-function \code{tree.lasso()} for tree-lasso}
\usage{
tree.lasso(
  x = x,
  y = y,
  lambda = 10,
  tree.parm = NULL,
  t.idx = FALSE,
  num.nonpen = 0,
  intercept = TRUE,
  mu = 0.01,
  threshold = 0,
  tol = 1e-06,
  NoVar = 50,
  y.mis = NULL,
  x.mis = NULL,
  t.glasso = FALSE,
  alpha = 1,
  gamma = 0,
  cov.proxy = "FL",
  L = 0,
  maxiter = 10000,
  predict.re = FALSE
)
}
\description{
Main function for tree-lasso
}
\details{
mixlasso
}
