% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.parms.R
\name{tree.parms}
\alias{tree.parms}
\title{Sub-function \code{tree.parms()} for tree-lasso}
\usage{
tree.parms(y = y, h = 0.7, impute = FALSE)
}
\description{
Tree structure from hierarchical cluster analysis for tree-lasso
}
\details{
mixlasso
}
