% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_surv_interval.R
\name{tune.surv.interval}
\alias{tune.surv.interval}
\title{Wrapper function for conditional logistic lasso objects.}
\usage{
tune.surv.interval(
  parms,
  x = x,
  y = y,
  x_test = NULL,
  y_test = NULL,
  family = "cox",
  strata.surv = NULL,
  intercept = TRUE,
  alpha = 0,
  lambda = NULL,
  num.nonpen = 0,
  nfolds = 3,
  foldid = NULL,
  seed = 12345,
  standardize.response = FALSE,
  p = NULL,
  type.measure = "deviance",
  type.min = "lambda.min",
  parallel = FALSE,
  verbose = FALSE,
  search.path = FALSE,
  ...
)
}
\arguments{
\item{x, y}{input matrix.}

\item{family}{response type.}

\item{strata.surv}{stratification variable for the Cox survival model.}

\item{intercept}{should  intercept(s) be fitted (default=\code{TRUE}) or set to zero (\code{FALSE}).}

\item{lambda}{optional user-supplied \code{lambda} sequence; default is NULL, and \code{espsgo} chooses its own sequence.}

\item{foldid}{an vector of values for the cross-validation.}

\item{seed}{random seed}

\item{standardize.response}{standardization for the response variables. Default: \code{TRUE}.}

\item{p}{the number of predictors from different data source.}

\item{parallel}{If \code{TRUE}, use parallel foreach to fit each fold except parallelizing each lambda for the tree-lasso methods. If \code{c(TRUE,TRUE)}, use parallel foreach to fit each fold and each lambda.}

\item{verbose}{print the middle search information, default is \code{TRUE}.}
}
\value{
An object with list "\code{tune.clogit.interval}"
\item{q.val}{the minimum MSE (or minus likelihood for the Cox model) through the cross-validation}
\item{model}{some model related quantities:
\itemize{\code{alpha  }}{the optimzed alpha}
\itemize{\code{lambda }}{the optimzed (first) lambda}
\itemize{\code{ipf    }}{the optimzed penalty factors}
\itemize{\code{p      }}{a vector of the numbers of features from multiple data sources}
\itemize{\code{nfolds }}{number of folds used for the cross-validation}
\itemize{\code{cvreg  }}{the cross-validation results}
}
}
\description{
Wrapper function for conditional logistic lasso objects optized based on the full likelihood by the EPSGO algorithm. This function is mainly used within the function \code{epsgo}.
}
\details{
mixlasso
}
