% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_tree_re_interval.R
\name{tune.tree.re.interval}
\alias{tune.tree.re.interval}
\title{Wrapper function for tree-lasso objects.}
\usage{
tune.tree.re.interval(
  parms,
  x,
  y,
  z = z,
  alpha = 1,
  nfolds = 5,
  foldid = NULL,
  cv.measure = cv.measure,
  tree.parm = tree.parm,
  num.nonpen = 0,
  seed = 12345,
  intercept = TRUE,
  standardize.response = FALSE,
  p = NULL,
  verbose = FALSE,
  parallel = FALSE,
  search.path = FALSE,
  threshold = threshold,
  tol = tol,
  mu = mu,
  NoVar = NoVar,
  y.mis = y.mis,
  x.mis = x.mis,
  t.idx = t.idx,
  t.glasso = t.glasso,
  maxiter = 10000,
  cov.proxy = "FL",
  L = 0,
  predict.re = predict.re,
  ...
)
}
\arguments{
\item{parms}{tuning parameter alpha for the tree-lasso object.}

\item{x, y}{\code{x} is a matrix where each row refers to a sample a each column refers to a gene; \code{y} is a factor which includes the class for each sample}

\item{nfolds}{number of cross-validation's folds, default 5.}

\item{foldid}{an optional vector of values between 1 and nfold identifying what fold each observation is in. If supplied, nfold can be missing.}

\item{num.nonpen}{number of predictors forced to be estimated (i.e., nonpenalization).}

\item{seed}{random seed}

\item{intercept}{should  intercept(s) be fitted (default=\code{TRUE}) or set to zero (\code{FALSE}).}

\item{standardize.response}{standardization for the response variables. Default: \code{TRUE}.}

\item{p}{the number of predictors from different data source.}

\item{verbose}{print the middle search information, default is \code{TRUE}.}

\item{parallel}{If \code{TRUE}, use parallel foreach to fit each lambda. If \code{c(TRUE,TRUE)}, use parallel foreach to fit each lambda and each fold.}

\item{search.path}{save the visited points, default is \code{FALSE}.}

\item{threshold}{threshold for estimated coefficients of the tree-lasso models.}

\item{lambda}{A user supplied lambda sequence.}
}
\description{
Wrapper function for tree-lasso objects used by epsgo function. This function is mainly used within the function \code{epsgo}.
}
\details{
mixlasso
}
